changes with files

+++ src/components/assets/UserProfile.js
@@ -137,8 +137,24 @@ class UserProfile extends Component {
 							})} className="btn common-btn-blue"><span>CANCEL</span></button>
 								<button className="btn common-btn-blue" onClick={((e)=>{
 										console.log(this.props.historyPush)
-										localStorage.clear();
-										this.props.historyPush.history.push("/Login");
+										let status;
+										fetch(Apiurl.LogoutCall.url,{
+											headers: {
+							                	"Content-Type" : "application/json",
+							                	"Authorization": 'Basic ' + localStorage.getItem("basic-auth"),
+							                },
+											method:Apiurl.LogoutCall.method
+										}).then(res=>{
+											status=res.status;
+											res.json()
+										}).then(data=>{
+											if(status===200){
+												localStorage.clear();
+												this.props.historyPush.history.push("/Login");
+											}else{
+												return false;
+											}
+										})
 								})}><span>YES</span></button>	
 							</div>
